New change

+++ src/main/java/com/irctc/model/Passenger.java
@@ -1,16 +1,14 @@
 package com.irctc.model;
 
-import java.math.BigDecimal;
+import java.math.BigInteger;
 
 import org.springframework.stereotype.Service;
 
 @Service
 public class Passenger {
-	private BigDecimal id;
+	private BigInteger id;
 	private String name;
 	private int age;
 	private char preference;
 	private int status;
-	
-
 }
